Renamed output adapters to interface adapters
